+++ redwood-treehouse-host/src/androidMain/kotlin/app/cash/redwood/treehouse/treehouseAppFactoryAndroid.kt
@@ -19,6 +19,7 @@ import android.content.Context
 import app.cash.redwood.leaks.LeakDetector
 import app.cash.zipline.loader.LoaderEventListener
 import app.cash.zipline.loader.ManifestVerifier
+import app.cash.zipline.loader.ZiplineHttpClient
 import app.cash.zipline.loader.asZiplineHttpClient
 import kotlinx.coroutines.Dispatchers
 import okhttp3.OkHttpClient
@@ -38,9 +39,36 @@ public fun TreehouseAppFactory(
   concurrentDownloads: Int = 8,
   stateStore: StateStore = MemoryStateStore(),
   leakDetector: LeakDetector = LeakDetector.none(),
+): TreehouseApp.Factory = TreehouseAppFactory(
+  context = context,
+  httpClient = httpClient.asZiplineHttpClient(),
+  manifestVerifier = manifestVerifier,
+  embeddedFileSystem = embeddedFileSystem,
+  embeddedDir = embeddedDir,
+  cacheName = cacheName,
+  cacheMaxSizeInBytes = cacheMaxSizeInBytes,
+  loaderEventListener = loaderEventListener,
+  concurrentDownloads = concurrentDownloads,
+  stateStore = stateStore,
+  leakDetector = leakDetector,
+)
+
+@Suppress("FunctionName")
+public fun TreehouseAppFactory(
+  context: Context,
+  httpClient: ZiplineHttpClient,
+  manifestVerifier: ManifestVerifier,
+  embeddedFileSystem: FileSystem? = null,
+  embeddedDir: Path? = null,
+  cacheName: String = "zipline",
+  cacheMaxSizeInBytes: Long = 50L * 1024L * 1024L,
+  loaderEventListener: LoaderEventListener = LoaderEventListener.None,
+  concurrentDownloads: Int = 8,
+  stateStore: StateStore = MemoryStateStore(),
+  leakDetector: LeakDetector = LeakDetector.none(),
 ): TreehouseApp.Factory = RealTreehouseApp.Factory(
   platform = AndroidTreehousePlatform(context),
-  httpClient = httpClient.asZiplineHttpClient(),
+  httpClient = httpClient,
   frameClockFactory = AndroidChoreographerFrameClock.Factory(),
   manifestVerifier = manifestVerifier,
   embeddedFileSystem = embeddedFileSystem,